Adjust body height for header
This patch adjust the content height to address the problem of the
overall height being too much, which would always cause scroll bars to
appear.

This also moves the Header code to a separate file in anticipation of
more code being added there later.
